#18 run.sh 파라미터 분기 도입
- typescript 빌드 도입 (STAND-ALONE 모드에서는 git pull 사용을 위해 src 파일을 제거하지 않음)

+++ run.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+build_mode=$1
+
 # 서버 내 .env 파일로 부터 환경변수 정의
 echo "---- set environment values ----"
 unamestr=$(uname)
@@ -12,23 +14,29 @@ elif [ "$unamestr"='FreeBSD' ] || [ "$unamestr"='Darwin' ]; then
 fi
 echo "---- OK ----"
 
-# 도커 이미지 빌드
-echo "---- container image build ----"
-build_cmd="docker build . -t wip-main-server-v2"
-while read line; do
-   arg_temp=$(echo $line | cut -f 1 -d'=')
-   build_cmd+=" --build-arg $arg_temp=$(eval echo '$'$arg_temp)"
-done < .env
+if [ $build_mode='STAND-ALONE' ]; then
+  tsc --build
+  npm install
+  node ./build/src/app.js
+elif [ $build_mode='SINGLE-CONTAINER' ]; then
+  # 도커 이미지 빌드
+  echo "---- container image build ----"
+  build_cmd="docker build . -t wip-main-server-v2"
+  while read line; do
+    arg_temp=$(echo $line | cut -f 1 -d'=')
+    build_cmd+=" --build-arg $arg_temp=$(eval echo '$'$arg_temp)"
+  done < .env
 
-$(echo $build_cmd)
-echo "---- OK ----"
+  $(echo $build_cmd)
+  echo "---- OK ----"
 
-# 컨테이너 종료. 실행중인 컨테이너가 있으면 강제로 제거
-echo "---- remove previous container ----"
-docker container rm -f wip-main-server-v2
-echo "---- OK ----"
+  # 컨테이너 종료. 실행중인 컨테이너가 있으면 강제로 제거
+  echo "---- remove previous container ----"
+  docker container rm -f wip-main-server-v2
+  echo "---- OK ----"
 
-# 컨테이너 실행
-echo "---- run container ----"
-docker run -d --name wip-main-server-v2 wip-main-server-v2
-echo "---- OK ----"
+  # 컨테이너 실행
+  echo "---- run container ----"
+  docker run -d --name wip-main-server-v2 wip-main-server-v2
+  echo "---- OK ----"
+fi
